pyproject.toml
[project]
name = "pyngldemos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyside6>=6.9.2",
    "numpy>=2.3.3",
    "qtpy>=2.4.3",
    "pyngl",
    "freetype-py>=2.5.1",
    "pysdl3>=0.9.8b9",
]


[tool.uv.workspace]
members=["*"]
exclude=["*"]
[tool.uv.sources]
pyngl = [
{ path = "/Users/jmacey/teaching/Code/PyNGL/", editable=true}
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
